requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Storage
supabase>=2.0.0
python-dotenv>=1.0.0

# UI
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6

# Scheduling
schedule>=1.2.0

# Hash and crypto
hashlib-additional>=1.0.0
